search action/reducer refactor
