+++ src/actions/entries.js
@@ -27,10 +27,6 @@ export const ENTRY_PERSIST_REQUEST = 'ENTRY_PERSIST_REQUEST';
 export const ENTRY_PERSIST_SUCCESS = 'ENTRY_PERSIST_SUCCESS';
 export const ENTRY_PERSIST_FAILURE = 'ENTRY_PERSIST_FAILURE';
 
-export const SEARCH_ENTRIES_REQUEST = 'SEARCH_ENTRIES_REQUEST';
-export const SEARCH_ENTRIES_SUCCESS = 'SEARCH_ENTRIES_SUCCESS';
-export const SEARCH_ENTRIES_FAILURE = 'SEARCH_ENTRIES_FAILURE';
-
 /*
  * Simple Action Creators (Internal)
  * We still need to export them for tests
@@ -122,35 +118,6 @@ export function emmptyDraftCreated(entry) {
     payload: entry,
   };
 }
-
-export function searchingEntries(searchTerm) {
-  return {
-    type: SEARCH_ENTRIES_REQUEST,
-    payload: { searchTerm },
-  };
-}
-
-export function searchSuccess(searchTerm, entries, page) {
-  return {
-    type: SEARCH_ENTRIES_SUCCESS,
-    payload: {
-      searchTerm,
-      entries,
-      page,
-    },
-  };
-}
-
-export function searchFailure(searchTerm, error) {
-  return {
-    type: SEARCH_ENTRIES_FAILURE,
-    payload: {
-      searchTerm,
-      error,
-    },
-  };
-}
-
 /*
  * Exported simple Action Creators
  */
@@ -242,25 +209,5 @@ export function persistEntry(collection, entryDraft) {
         }));
         dispatch(entryPersistFail(collection, entry, error));
       });
-  };
-}
-
-export function searchEntries(searchTerm, page = 0) {
-  return (dispatch, getState) => {
-    const state = getState();
-    let collections = state.collections.keySeq().toArray();
-    collections = collections.filter(collection => selectIntegration(state, collection, 'search'));
-    const integration = selectIntegration(state, collections[0], 'search');
-    if (!integration) {
-      dispatch(searchFailure(searchTerm, 'Search integration is not configured.'));
-    }
-    const provider = integration ?
-      getIntegrationProvider(state.integrations, integration)
-      : currentBackend(state.config);
-    dispatch(searchingEntries(searchTerm));
-    provider.search(collections, searchTerm, page).then(
-      response => dispatch(searchSuccess(searchTerm, response.entries, response.pagination)),
-      error => dispatch(searchFailure(searchTerm, error))
-    );
   };
 }
