+++ src/containers/SearchPage.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import { connect } from 'react-redux';
 import { selectSearchedEntries } from '../reducers';
-import { searchEntries } from '../actions/entries';
+import { searchEntries } from '../actions/search';
 import { Loader } from '../components/UI';
 import EntryListing from '../components/EntryListing/EntryListing';
 import styles from './breakpoints.css';
@@ -32,7 +32,7 @@ class SearchPage extends React.Component {
 
   handleLoadMore = (page) => {
     const { searchTerm, searchEntries } = this.props;
-    searchEntries(searchTerm, page);
+    if (!isNaN(page)) searchEntries(searchTerm, page);
   };
 
   render() {
